requirements.txt
loguru
watchdog
pyyaml
dictor